==> final_project/02_creating_ru_vectors.ipynb <==
from gensim.models.fasttext import FastText
import fasttext
import numpy as np
import matplotlib.pyplot as plt
import nltk
from string import punctuation
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from pymorphy2 import MorphAnalyzer
morph = MorphAnalyzer()
from nltk.tokenize import sent_tokenize
from nltk import WordPunctTokenizer
import pandas as pd

import nltk
nltk.download('stopwords')

import re

data = pd.read_csv('D:\\python\\CL\\gob\\pi222.csv', sep=';', low_memory=False)
data = data[['Russian', 'English']]
index_names = data[data['Russian'].str.contains("тест" or "test")==True].index
data.drop(index_names, inplace = True)
data = data.dropna()
data = data[data.Russian != data.English]
data = data.drop_duplicates(subset=['English', 'Russian'])
data_ru = data['Russian'].tolist()

def preprocess_text(item):
    item = re.sub(r"\d+%", " ", item)
    item = re.sub(r"x\d+", " ", item)
    item = re.sub(r"\d+", " ", item)
    item = re.sub(r"\n", " ", item)
    item = re.sub(r"\[.+\]", " ", item)
    item = re.sub(r"\\.+\\;", " ", item)
    item = re.sub(r"http.+", " ", item)
    item = re.sub(r"\{.*\}", " ", item)
    item = re.sub(r" [xX] ", " ", item)
    item = re.sub(r"%[sd]", " ", item)
    item = re.sub(r"<.+>", " ", item)
    item = re.sub(r"[\U00010000-\U0010ffff]", " ", item)
    item = re.sub(r"[!@#$%\^\&\*()_=+\?\!:;\",\.\\»«—№]", " ", item)
    item = re.sub(r"\s+", " ", item)
    item = item.strip(' ')
    item = item.lower()
    
    tokens = item.split()
    tokens = [morph.parse(word)[0].normal_form for word in tokens]
    tokens = [word for word in tokens if word not in stopwords.words('russian')]
    preprocessed_text = ' '.join(tokens)

    return preprocessed_text

final_corpus = [preprocess_text(sentence) for sentence in data_ru if sentence.strip() !='']

word_punctuation_tokenizer = nltk.WordPunctTokenizer()
word_tokenized_corpus = [word_punctuation_tokenizer.tokenize(sent) for sent in final_corpus]

embedding_size = 15
window_size = 60
min_word = 5
down_sampling = 1e-2

ft_model = FastText(word_tokenized_corpus,
                      size=embedding_size,
                      window=window_size,
                      min_count=min_word,
                      sample=down_sampling,
                      sg=1,
                      iter=100)

print(ft_model.wv['ключ'])

semantically_similar_words = {words: [item[0] for item in ft_model.wv.most_similar([words], topn=5)]
                  for words in ['кристалл', 'наоми', 'ключ', 'остров', 'строить', 'награда']}

for k,v in semantically_similar_words.items():
    print(k+":"+str(v))

print(ft_model.wv.similarity(w1='мано', w2='дорин'))
print(ft_model.wv.similarity(w1='отель', w2='событие'))
print(ft_model.wv.similarity(w1='ключ', w2='сундук'))

from gensim.models.fasttext import save_facebook_model
save_facebook_model(ft_model, "ru_model_fb.bin", encoding='utf-8')
