==> final_project/05_CountVector.ipynb <==
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.feature_extraction import text
import pandas as pd
import re
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer
import numpy as np  
stemmer = WordNetLemmatizer()
import spacy
from nltk.corpus import stopwords
import torch
import tensorflow
from sklearn.metrics.pairwise import cosine_similarity
import string 
import random
import nltk

import spacy
nlp = spacy.load('en_core_web_sm')
nlp.max_length = 2100000

def preprocess_text(item):
    item = re.sub(r"\d+%", " ", item)
    item = re.sub(r"x\d+", " ", item)
    item = re.sub(r"\d+", " ", item)
    item = re.sub(r"\n", " ", item)
    item = re.sub(r"\[.+\]", " ", item)
    item = re.sub(r"\\+.+;", " ", item)
    item = re.sub(r"http.+", " ", item)
    item = re.sub(r"\{.*\}", " ", item)
    item = re.sub(r" [xX] ", " ", item)
    item = re.sub(r"%[sd]", " ", item)
    item = re.sub(r"<.+>", " ", item)
    item = re.sub(r"[\U00010000-\U0010ffff]", " ", item)
    item = re.sub(r"[!@#$%\^\&\*()_=+\?\!:;\",\.\\»«—✨]", " ", item)
    item = re.sub(r"\s+", " ", item)
    item = item.strip(' ')
    item = item.lower()
    
    tokens = item.split()
    tokens = [nlp(word)[0].lemma_ if word != "flowerbed" else "flowerbed" for word in tokens]
    tokens = [word for word in tokens if word not in stopwords.words('english')]
    preprocessed_text = ' '.join(tokens)

    return preprocessed_text

data = pd.read_csv('pi2.csv', sep=';', low_memory=False)
data = data[['string_id','Russian', 'English']]
index_names = data[data['Russian'].str.contains("тест" or "test")==True].index
data.drop(index_names, inplace = True)
data = data.dropna()
data = data.drop_duplicates()
data = data[data.Russian != data.English]
index_names2 = data[data['string_id'].str.contains("achieve", "title") ==True].index
data.drop(index_names2, inplace = True)
data = data[['Russian', 'English']].drop_duplicates()
data_en = data['English'].tolist()
final_corpus = [preprocess_text(sentence) for sentence in data_en if len(sentence) <= 80 and len(sentence) > 0]

data_str = ' '.join(final_corpus)

nlp = spacy.load('en_core_web_sm')
nlp.max_length = 2100000
doc = nlp(data_str)

nouns = set()
for token in doc:
    if token.pos_ == "NOUN":
        nouns.add(token.lemma_)

verbs = set()
for token in doc:
    if token.pos_ == "VERB":
        if token.text != "\'ve" and token.text != "s" and token.text != "\'m" and token.text != "\'re":
            if len(token.text) > 2:
                verbs.add(token.lemma_)

all_pos = nouns.union(verbs)

stop_words = "english"

# Extract candidate words/phrases
count = TfidfVectorizer(min_df = 5, stop_words=stop_words).fit(final_corpus)
bag_of_words = count.transform(final_corpus)
sum_words = bag_of_words.sum(axis=0)
words_freq = [(word, sum_words[0, idx]) for word, idx in count.vocabulary_.items()]
words_freq =sorted(words_freq, key = lambda x: x[1], reverse=True)
all_candidates = []
for item in words_freq[:700]:
    all_candidates.append(item[0])

candidates = [x for x in all_candidates in x in nouns]
candidates2 = [x for x in all_candidates in x in verbs]
keywords = set(candidates + candidates2)

print(keywords)

import pickle
with open('keywords_short.pkl', 'wb') as f:
       pickle.dump(set(keywords), f)
