==> final_project/06_TermBase.ipynb <==
#составление переводного глоссария

import pandas as pd
import re
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer
import numpy as np  
stemmer = WordNetLemmatizer()
from string import punctuation
import numpy

punctuation = punctuation.replace('\"', "")

data = pd.read_csv('pi2.csv', sep=';', low_memory=False)
data = data[['string_id','Russian', 'English']]
index_names = data[data['Russian'].str.contains("тест" or "test")==True].index
# # index_names
data.drop(index_names, inplace = True)
data = data.dropna()
data = data.drop_duplicates()
data = data[data.Russian != data.English]
index_names2 = data[data['string_id'].str.contains("achieve", "title") ==True].index
data.drop(index_names2, inplace = True)
data = data[data['English'].apply(lambda x: len(x.split()) <= 3)]
data = data.drop_duplicates(subset=['English', 'Russian'], keep='last')
data = data[['Russian', 'English']]

import pickle
with open('keywords.pkl', 'rb') as f:
    keywords = pickle.load(f)

data_dict = dict(zip(data.English,data.Russian))

termbase = {}
for x in data_dict:
    if x.lower().strip(punctuation) in keywords:
        termbase[x.strip(punctuation)] = data_dict[x].strip(punctuation)

termbase_df = pd.DataFrame.from_dict(termbase, orient='index')
termbase_df.drop_duplicates()

with open('D:\\python\\CL\\alignment\\termbase_unigrams_X.pkl', 'rb') as f:
    unigrams = pickle.load(f)
with open('D:\\python\\CL\\alignment\\termbase_3_X.pkl', 'rb') as f:
    unigrams3 = pickle.load(f)

termbase_en_lower = [x.lower() for x in termbase]
for x in unigrams:
    if x not in termbase_en_lower:
        termbase[x] = unigrams[x]
print(termbase)         

#поиск лучших соответствий по топ-3 ближайшим словам
def options():
    print("Какой термин вы ищете?")
    term = input()
    with open('D:\\python\\CL\\alignment\\termbase_3_X.pkl', 'rb') as f:
        unigrams3 = pickle.load(f)
    print(unigrams3[term])

options()

#проверим сколько у нас совпадений со сделанной вручную базой
df_tb = pd.read_csv('PI2TB.csv', sep=',')
df_tb = df_tb.English.drop_duplicates()
list_tb = df_tb.tolist()
list_tb_clean = [x.lower().strip() for x in list_tb]
len(list_tb_clean)

our_tb_clean = [x.lower().strip() for x in list(termbase.keys())]
common = [x for x in list_tb_clean if x in our_tb_clean]
print(len(common), common)

termbase_df.to_csv("termbase1.csv", encoding = "utf-8")
